# backend/requirement/api/docs/swagger.yaml
openapi: 3.0.0
info:
  title: AI Task Dashboard API
  description: >
    Automatically derived CRUD API documentation based on `Data requirements.md`.
    All responses follow the standardized API response envelope.
  version: 1.1.0
servers:
  - url: /api/v1
    description: Main API Server
...
paths:
...
  /activities:
    get:
      tags: [Activities]
      summary: Get all activities
      description: Highly dynamic list, heavily utilizes pagination limits and offsets.
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: A list of activities including contextual derived fields
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          activities:
                            type: array
                            items:
                              $ref: '#/components/schemas/ActivityResponse'
    post:
      tags: [Activities]
      summary: Create an activity
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityBase'
      responses:
        '201':
          description: Activity created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          activity:
                            $ref: '#/components/schemas/ActivityResponse'
...
  /activities/{id}:
    get:
      tags: [Activities]
      summary: Get activity by MongoDB _id
      parameters:
        - $ref: '#/components/parameters/MongoIdParam'
      responses:
        '200':
          description: Activity object with derived fields
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          activity:
                            $ref: '#/components/schemas/ActivityResponse'
    put:
      tags: [Activities]
      summary: Update an activity
      parameters:
        - $ref: '#/components/parameters/MongoIdParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityBase'
      responses:
        '200':
          description: Activity updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          activity:
                            $ref: '#/components/schemas/ActivityResponse'
    delete:
      tags: [Activities]
      summary: Delete an activity
      parameters:
        - $ref: '#/components/parameters/MongoIdParam'
      responses:
        '204':
          description: Activity deleted
  /activities/user/{userId}:
    get:
      tags: [Activities]
      summary: Get activities by user ID
      operationId: getActivitiesByUserId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: 24-character hexadecimal MongoDB User ID
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: A list of activities for the specified user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          activities:
                            type: array
                            items:
                              $ref: '#/components/schemas/ActivityResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEnvelope'
  /activities/user/{userId}/count:
    get:
      tags: [Activities]
      summary: Get total activity count for a user
      operationId: getActivityCountByUserId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: 24-character hexadecimal MongoDB User ID
      responses:
        '200':
          description: Total activity count for the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountApiResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEnvelope'

  # ==============================
  # NOTIFICATIONS ENDPOINTS
  # ==============================
...
components:
  parameters:
    MongoIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: 24-character hexadecimal MongoDB Object ID
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 20
      description: Amount of records to return for pagination
    OffsetParam:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        default: 0
      description: Amount of records to skip for pagination
  schemas:
...
    ApiResponseEnvelope:
      type: object
      required: [status, message, timestamp]
      properties:
        status:
          type: string
          enum: [success, failed]
          description: Indicates whether the request was successful or failed
        message:
          type: string
          description: Human-readable success or failure message
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 date-time of the response
        data:
          type: object
          nullable: true
          description: Domain-specific payload. May be null on failure.
    CountApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                count:
                  type: integer
                  description: The total count

    # ------------------
    # MINIMAL SCHEMAS (FOR REUSE)
    # ------------------
    UserMinimal:
      type: object
      required: [_id, name]
      properties:
        _id:
          type: string
          description: MongoDB hex ID
        name:
          type: string
        email:
          type: string
        companyTechRole:
          type: string
        role:
          type: string
          enum: ['ADMIN', 'MEMBER', 'USER']
    ChallengeMinimal:
      type: object
      required: [_id, title]
      properties:
        _id:
          type: string
          description: MongoDB hex ID
        virtualId:
          type: string
          description: CH-XXX format
        title:
          type: string
    IdeaMinimal:
      type: object
      required: [_id, title]
      properties:
        _id:
          type: string
          description: MongoDB hex ID
        ideaId:
          type: string
          description: ID-XXXX format
        title:
          type: string
...
    ActivityBase:
      type: object
      properties:
        type:
          type: string
          enum:
            - challenge_created
            - idea_created
            - challenge_status_update
            - challenge_edited
            - idea_edited
            - challenge_upvoted
            - idea_upvoted
            - challenge_commented
            - idea_commented
            - challenge_subscribed
            - idea_subscribed
            - challenge_deleted
            - idea_deleted
            - log_in
            - log_out
        fk_id:
          type: string
          nullable: true
        userId:
          type: string
        month:
          type: integer
        year:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    ActivityResponse:
      allOf:
        - $ref: '#/components/schemas/ActivityBase'
        - type: object
          properties:
            _id:
              type: string
            userDetails:
              $ref: '#/components/schemas/UserMinimal'
            entityDetails:
              type: object
              oneOf:
                - $ref: '#/components/schemas/ChallengeMinimal'
                - $ref: '#/components/schemas/IdeaMinimal'
              description: "(DERIVED) What was acted upon (Challenge or Idea)"
            activitySummary:
              type: string
              description: "(DERIVED) Human readable string generated dynamically"
...